Crate plc_lowering

Modules

inheritance

This module is responsible for lowering inherited references. Derived POUs can access members of their base class using the __SUPER qualifier, which is an implicit reference to an instance of the base class which is added as a member variable of the derived class in the ‘pre_index’ pass.

initializer

The initializer lowering module generates constructor functions (__ctor) for every user-defined type and POU in a compilation unit. It traverses the AST to collect initialization logic and injects it as new POU definitions before code generation.