plc_lowering

Module control_statement

Control Statement Lowering

This module handles the lowering of control statements in an effort to ensure that the control statements are processed without any side-effects that the user might not expect.

Current use-cases for lowering control statements

1. else/if statements

Due to the existing way that the AST was lowered, if a function call returned an aggregate type and that function call was part of the condition of an else/if, then that function would always be evaluated. This would result in some cases where an operation in the function could process when the user never intended it to.

For example:

FUNCTION foo : STRING[10]
    VAR
        result : STRING[10];
    END_VAR
    VAR_IN_OUT
        counter : INT;
    END_VAR
    result := 'Hello';
    counter := counter + 1;
    foo := result;
END_FUNCTION

FUNCTION main
    VAR
        counterVar: INT;
    END_VAR

    counterVar := 0;

    IF foo(counterVar) = 'Hello' THEN
        // Do nothing
    ELSIF foo(counterVar) = 'Goodbye' THEN
        // Do nothing
    END_IF
END_FUNCTION

In the above example, the foo() method would be called a second time and increment the counter variable to 2. This is obviously incorrect as the statement evaluation should have ended after the IF returned true.

Thus, the statement is now lowered to prevent this evaluation from occuring, the resulting lowered code would look like this:

    ...
    IF foo(counterVar) = 'Hello' THEN
        // Do nothing
    ELSE
        IF foo(counterVar) = 'Goodbye' THEN
            // Do nothing
        END_IF
    END_IF
    ...

Structs

ControlStatementLowerer

ControlStatementParticipant