plc_lowering

Module reference_to_return

Reference To Return Lowering

This module handles the lowering of functions that return a “REFERENCE TO” to hide additional processing that creates a temporary variable and assigns the result to that instead so that no values are assigned directly from the stack.

The lowering mechanism works as follows:

Functions with a REFERENCE TO return type will be lowered to a VOID return type.

Functions with a REFERENCE TO return type will be given a new VAR_INPUT variable: __{function_name}_return_val that any return value of the function will be assigned to.

Any POU that invokes a function with a REFERENCE TO return type will receive a new temporary variable that is then passed as the first parameter to the function.

Any POU that invokes a function with a REFERENCE TO return type will then lower the call to that function from:

    refVal REF= exampleFunc(refVal);

to the following:

    __exampleFunc_return_val_1 REF= __exampleFunc_return_val_store_1;
    exampleFunc(__exampleFunc_return_val_1, refVal);
    refVal REF= __exampleFunc_return_val_1;

Structs

ReferenceToReturnImplementationCallContext

ReferenceToReturnImplementationContext

ReferenceToReturnImplementationContextCalls

ReferenceToReturnImplementationPouContext

ReferenceToReturnLowerer

ReferenceToReturnParticipant

ReferenceToReturnPouContext